Initial support for QHY9 as an INDI driver
This project uses libqhyccd (QHY Linux SDK) to implement a QHY9 driver.
This should be the basis for supporting QHY cameras under INDI. I don't
believe that it makes sense for the maintainer of the INDI driver to have
to worry about talking to the camera, etc. This all can be handled by the
manufacturer - QHY - in their Linux SDK (libqhyccd). Furthermore, it would
appear that it may be a lot easier to support all QHY cameras with their
new SDK, wo this project could easily become indi-qhy ;-)

What is supported:
  - Exposures at 1x1 bin (Light only - no special shutter control)
  - TEC temperature control

What doesn't work:
  - Subframing
  - Binning
  - Special shutter control (i.e. darks)
  - Controls for gain, offset, etc.

NOTE: For now, this project requires the use of my fork of the QHY library.
I know this contradicts the paragraph above, I have already submitted a pull
request and hopefully this won't be the case for much longer.